requirements.txt
pandas
python-dotenv
requests
prometheus_client